+++ spotinst/resource_spotinst_ocean_aws_test.go
@@ -614,7 +614,7 @@ const testStrategyConfig_EmptyFields = `
 // region OceanAWS: Scheduling
 func TestAccSpotinstOceanAWS_Scheduling(t *testing.T) {
 	clusterName := "test-acc-cluster-scheduling"
-	controllerClusterID := "scheduling-controller-id"
+	controllerClusterID := "NoamCSTFTest"
 	resourceName := createOceanAWSResourceName(clusterName)
 
 	var cluster aws.Cluster
@@ -721,7 +721,7 @@ const testSchedulingConfig_EmptyFields = `
 // region OceanAWS: Autoscaler
 func TestAccSpotinstOceanAWS_Autoscaler(t *testing.T) {
 	clusterName := "test-acc-cluster-autoscaler"
-	controllerClusterID := "autoscaler-controller-id"
+	controllerClusterID := "NoamCSTFTest"
 	resourceName := createOceanAWSResourceName(clusterName)
 
 	var cluster aws.Cluster
@@ -743,6 +743,7 @@ func TestAccSpotinstOceanAWS_Autoscaler(t *testing.T) {
 					testCheckOceanAWSAttributes(&cluster, clusterName),
 					resource.TestCheckResourceAttr(resourceName, "autoscaler.#", "1"),
 					resource.TestCheckResourceAttr(resourceName, "autoscaler.0.autoscale_cooldown", "300"),
+					resource.TestCheckResourceAttr(resourceName, "autoscaler.0.auto_headroom_percentage", "30"),
 					resource.TestCheckResourceAttr(resourceName, "autoscaler.0.autoscale_down.#", "1"),
 					resource.TestCheckResourceAttr(resourceName, "autoscaler.0.autoscale_down.0.evaluation_periods", "300"),
 					resource.TestCheckResourceAttr(resourceName, "autoscaler.0.autoscale_down.0.max_scale_down_percentage", "50"),
@@ -770,6 +771,7 @@ func TestAccSpotinstOceanAWS_Autoscaler(t *testing.T) {
 					testCheckOceanAWSAttributes(&cluster, clusterName),
 					resource.TestCheckResourceAttr(resourceName, "autoscaler.#", "1"),
 					resource.TestCheckResourceAttr(resourceName, "autoscaler.0.autoscale_cooldown", "600"),
+					resource.TestCheckResourceAttr(resourceName, "autoscaler.0.auto_headroom_percentage", "60"),
 					resource.TestCheckResourceAttr(resourceName, "autoscaler.0.autoscale_down.#", "1"),
 					resource.TestCheckResourceAttr(resourceName, "autoscaler.0.autoscale_down.0.evaluation_periods", "600"),
 					resource.TestCheckResourceAttr(resourceName, "autoscaler.0.autoscale_down.0.max_scale_down_percentage", "10"),
@@ -820,6 +822,7 @@ const testScalingConfig_Create = `
     autoscale_is_enabled     = false
     autoscale_is_auto_config = false
     autoscale_cooldown       = 300
+    auto_headroom_percentage = 30
 
     autoscale_headroom = {
       cpu_per_unit    = 1024
@@ -849,6 +852,7 @@ const testScalingConfig_Update = `
     autoscale_is_enabled     = true
     autoscale_is_auto_config = true
     autoscale_cooldown       = 600
+    auto_headroom_percentage = 60
 
     autoscale_headroom = {
       cpu_per_unit    = 512
@@ -902,7 +906,7 @@ const testScalingConfig_EmptyFields = `
 
 func TestAccSpotinstOceanAWS_UpdatePolicy(t *testing.T) {
 	clusterName := "test-acc-cluster-update-policy"
-	controllerClusterID := "update-policy-controller-id"
+	controllerClusterID := "NoamCSTFTest"
 	resourceName := createOceanAWSResourceName(clusterName)
 
 	var cluster aws.Cluster