+++ spotinst/resource_spotinst_ocean_gke_import.go
@@ -3,6 +3,7 @@ package spotinst
 import (
 	"context"
 	"fmt"
+	"github.com/terraform-providers/terraform-provider-spotinst/spotinst/ocean_gke_import_scheduling"
 	"log"
 	"time"
 
@@ -35,7 +36,7 @@ func setupClusterGKEImportResource() {
 	fieldsMap := make(map[commons.FieldName]*commons.GenericField)
 
 	ocean_gke_import.Setup(fieldsMap)
-	//ocean_gke_import_scheduling.Setup(fieldsMap)
+	ocean_gke_import_scheduling.Setup(fieldsMap)
 
 	commons.OceanGKEImportResource = commons.NewOceanGKEImportResource(fieldsMap)
 }